umount-usb.sh: created functions;checks if USB connected/mounted.

+++ umount-usb.sh
@@ -7,8 +7,7 @@
 # Author       : Copyright © 2025, Richard B. Romig, Mosfanet
 # Email        : rick.romig@gmail.com | rick.romig@mymetronet.com
 # Created      : 12 Mar 2025
-# Last updated : 16 Sep 2025
-# Version      : 1.2.25259
+# Last updated : 05 Dec 2025
 # Comments     : Only unmounts devices belonging to the current user.
 # TODO (Rick)  :
 # License      : GNU General Public License, version 2.0
@@ -25,10 +24,40 @@
 # GNU General Public License for more details.
 ##########################################################################
 
-user=$(whoami)
-usb_drive=$(find /media/"$user" -maxdepth 1 -type d -user "$user" | fzf --height 40% --reverse --prompt "Select the USB drive to unmount: ")
+## Shellcheck Directives ##
+# shellcheck source=/home/rick/bin/functionlib
 
-umount -l "$usb_drive"
-printf "%s unmounted.\n" "$usb_drive"
-[[ -d "$usb_drive" ]] && rmdir "$usb_drive"
-exit
+# Load function library
+
+if [[ -x "$HOME/bin/functionlib" ]]; then
+  source "$HOME/bin/functionlib"
+else
+  printf "\e[91mERROR:\e[0m functionlib not found!\n" >&2
+  exit 1
+fi
+
+check_usb() {
+	[[ $(grep 'usb' <(lsblk -S -o TRAN)) = *usb* ]] && return 0 || return 1
+}
+
+unmount_usb() {
+	local user usb_drive
+	check_usb || { printf "No USB drive connected!\n" >&2; return; }
+	user=$(whoami)
+	usb_drive=$(find /media/"$user" -maxdepth 1 -type d -user "$user" | fzf --height 40% --reverse --prompt "Select the USB drive to unmount: ")
+	[[ "$usb_drive" ]] || { printf "No USB drives mounted!\n" >&2; return; }
+	umount -l "$usb_drive"
+	printf "\"%s\" unmounted or selected.\n" "$usb_drive"
+	[[ -d "$usb_drive" ]] && rmdir "$usb_drive"
+}
+
+main() {
+	local script="${0##*/}"
+	local version="2.0.25339"
+	check_package fzf
+	unmount_usb
+	over_line "$script $version"
+	exit
+}
+
+main "$@"
